// src/components/TaskHistory/interface.ts
/**
 * Интерфейс, описывающий событие в графе истории задачи
 */
export interface TaskHistoryBlockInfo {
	/**
	 * Тип события (успех, ошибка, предупреждение, информация)
	 */
	type: "success" | "error" | "warning" | "info";
	/**
	 * Текст сообщения
	 */
	text?: string;
	/**
	 * Принятое решение
	 */
	decision?: {
		/**
		 * Роль пользователя, принявшего решение
		 */
		by: "operator" | "controller";
		/**
		 * Статус: успех/ошибка
		 */
		status: boolean;
		/**
		 * Пользователь
		 */
		user: {
			/**
			 * Полное имя пользователя
			 */
			name: string;
		};
		/**
		 * Дата принятия решения
		 */
		date: string;
		/**
		 * Текст решения
		 */
		label: string;
	};
}